Create Version 2.0.0
In these updates, we implemented a hybrid encryption model in the backend, which includes RSA and AES encryption. Additionally, we designed a custom standard for rearranging file characters in an encryption pattern to ensure that user data is not easily accessible and remains protected from disclosure.

In the application's user interface, file encryption and decryption features have been implemented, several bugs have been fixed, and new features have been added, such as file path selection in the settings. Improvements have also been made in error handling, as well as updates in the history section and management of simultaneous encryption and decryption of text and files.v